test(react): add test for wrong state being dispatched
